move activity-config into src/config folder
